Define a global exception handler.
